Refactor | Oauth Namespace
