+++ composer.json
@@ -25,7 +25,8 @@
             "database/factories"
         ],
         "psr-4": {
-            "App\\": "app/"
+            "App\\": "app/",
+            "Oauth\\": "oauth/"
         }
     },
     "autoload-dev": {
